- now user data is joined before sending the posts - cleaned up main.py - added realtion to modle.Post

+++ apps/routers/posts.py
@@ -1,4 +1,4 @@
-from typing import List
+from typing import List, Optional
 from fastapi import APIRouter, Depends, Response, status, HTTPException
 from sqlalchemy.orm import Session
 from apps import schemas
@@ -10,16 +10,20 @@ from ..database import get_db
 router = APIRouter(prefix="/posts", tags=['posts'])
 
 @router.post('/', status_code=status.HTTP_201_CREATED, response_model=schemas.Post)
-async def create_post(posts: schemas.PostBase, db: Session = Depends(get_db), current_user: str = Depends(get_current_user)):
+async def create_post(posts: schemas.PostCreate, db: Session = Depends(get_db), current_user: models.Users = Depends(get_current_user)):
     new_post = models.Post(**posts.model_dump())
+    new_post.user_id = current_user.id
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
     return new_post
 
+@router.get('/', response_model=List[schemas.Post])
+async def get_all_posts(db: Session = Depends(get_db), limit: int = 10, skip: int = 0, search: Optional[str] = ""):
+    return db.query(models.Post).filter(models.Post.title.contains(search)).limit(limit=limit).offset(skip)
 
 @router.get('/{id}', response_model=schemas.Post)
-async def get_post(id: int, response: Response, db: Session = Depends(get_db)):
+async def get_post(id: int, response: Response, db: Session = Depends(get_db), current_user: models.Users = Depends(get_current_user)):
     # cursor.execute("""SELECT * FROM posts WHERE id = %s """, (str(id),))
     # post = cursor.fetchone()
     post = db.query(models.Post).where(models.Post.id == id).first()
@@ -29,18 +33,20 @@ async def get_post(id: int, response: Response, db: Session = Depends(get_db)):
 
 
 @router.delete('/{id}', status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(id: int, response: Response, db: Session = Depends(get_db)):
+async def delete_post(id: int, response: Response, db: Session = Depends(get_db), current_user: models.Users = Depends(get_current_user)):
     # cursor.execute("""DELETE FROM posts where id = %s RETURNING *""", (str(id),))
     # delete_post = cursor.fetchone()
     query = db.query(models.Post).where(models.Post.id == id)
     if not query.first():
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="The given id couldn't be found!")
+    if query.first().user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="This post was not created by you.")
     query.delete()
     db.commit()
     return delete_post
     
 @router.put('/{id}', status_code=status.HTTP_200_OK, response_model=schemas.Post)
-async def update_post(id: int, post: PostBase, db: Session = Depends(get_db)):
+async def update_post(id: int, post: PostBase, db: Session = Depends(get_db), current_user: models.Users = Depends(get_current_user)):
     # cursor.execute("""UPDATE posts SET title = %s, content = %s, published = %s WHERE id = %s RETURNING *""", 
     #                (str(post.title), str(post.content), str(post.published), str(id),))
     # updated_post = cursor.fetchone()
@@ -50,6 +56,8 @@ async def update_post(id: int, post: PostBase, db: Session = Depends(get_db)):
     if not query.first():
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"The said id {id} doesn't exist")
     else:
+        if query.first().user_id != current_user.id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="This post was not created by you.")
         query.update(post.model_dump())
         db.commit()
         return db.query(models.Post).where(models.Post.id == id).first()
